+++ RELEASE-NOTES.md
@@ -5,7 +5,7 @@
 * #61 Typo in JobRegistry.allowsSimlutaneousExecutions() method name 
 * #62 Remove circular dependency DefaultJobRegistry <-> Scheduler
 * #64 Consul-based job locks
-* #65 Changed cron type from String to value object.
+* #65 Value object: Cron
 * #66 Switch TriggerDescriptor properties to value objects
 * #72 Double --job option
 * #76 Scheduled job exceptions are not reported