+++ RELEASE-NOTES.md
@@ -7,6 +7,7 @@
 * #64 Consul-based job locks
 * #65 Value object: Cron
 * #66 Switch TriggerDescriptor properties to value objects
+* #71 ConsulLockHandler local job locking
 * #72 Double --job option
 * #76 Scheduled job exceptions are not reported